+++ src/handleable.rs
@@ -33,11 +33,12 @@ where
   /// Handle an `Event`.
   ///
   /// The widget has the option to either consume the event and return
-  /// nothing, in which no one else will get informed about it, forward
-  /// it directly (the default behavior), in which case the its parent
-  /// widget will receive it, or return a completely different event.
+  /// nothing, in which case no one else will get informed about it,
+  /// forward it directly (the default behavior), in which case its
+  /// parent widget will receive it, or return a completely different
+  /// event.
   #[allow(unused_variables)]
-  fn handle(&mut self, cap: &mut dyn MutCap<E>, event: E) -> Option<UiEvents<E>> {
+  fn handle(&self, cap: &mut dyn MutCap<E>, event: E) -> Option<UiEvents<E>> {
     // By default we just pass through the event, which will cause it to
     // bubble up to the parent.
     Some(event.into())
@@ -45,17 +46,13 @@ where
 
   /// Handle a custom event.
   #[allow(unused_variables)]
-  fn handle_custom(&mut self, cap: &mut dyn MutCap<E>, event: Box<dyn Any>) -> Option<UiEvents<E>> {
+  fn handle_custom(&self, cap: &mut dyn MutCap<E>, event: Box<dyn Any>) -> Option<UiEvents<E>> {
     Some(UiEvent::Custom(event).into())
   }
 
   /// Handle a custom event without transferring ownership of it.
   #[allow(unused_variables)]
-  fn handle_custom_ref(
-    &mut self,
-    cap: &mut dyn MutCap<E>,
-    event: &mut dyn Any,
-  ) -> Option<UiEvents<E>> {
+  fn handle_custom_ref(&self, cap: &mut dyn MutCap<E>, event: &mut dyn Any) -> Option<UiEvents<E>> {
     None
   }
 }